refactor: 동일한 책임을 수행하는 객체들의 추상화를 위한 인터페이스 생성
생성된 인터페이스
- RacingVehicle 인터페이스. race에 대한 메시지를 수신하여 Report를 반환하는 책임
- Report 인터페이스. 문자열 형태의 기록 결과를 반환할 책임
