+++ src/main/java/racingcar/model/racingvehicle/vehicle/RacingVehicle.java
@@ -1,0 +1,7 @@
+package racingcar.model.racingvehicle.vehicle;
+
+import racingcar.model.report.Report;
+
+public interface RacingVehicle {
+    Report race();
+}
