+++ src/main/java/racingcar/model/report/Report.java
@@ -1,0 +1,5 @@
+package racingcar.model.report;
+
+public interface Report {
+    String report();
+}
